requirements.txt
pandas
numpy
matplotlib
seaborn
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
python-box
pyYAML
tqdm
ensure
notebook
openpyxl
dill
scikit-learn
flask-restplus
flask-restx
azure-storage-blob
-e .